+++ muk_utils/models/res_config_settings.py
@@ -36,11 +36,11 @@ class ResConfigSettings(models.TransientModel):
     #----------------------------------------------------------
     
     attachment_location = fields.Selection(
-        selection=_attachment_location_selection,
+        selection=lambda self: self._attachment_location_selection(),
         string='Storage Location',
         required=True,
+        default='file',
         help="Attachment storage location.")
-    
     
     attachment_location_changed = fields.Boolean(
         compute='_compute_attachment_location_changed',