+++ muk_utils/doc/changelog.rst
@@ -1,3 +1,13 @@
+`1.6.0`
+-------
+
+- Override Attachment to make it more extendable
+
+`1.5.0`
+-------
+
+- Storage Migration Action
+
 `1.4.0`
 -------
 
@@ -16,9 +26,9 @@
 `1.1.0`
 -------
 
-- Added storage settings
+- Added Storage Settings
 
 `1.0.0`
 -------
 
-- Init version
+- Init Version